Refactor cabinet Telegram auth into reusable components
